Moving construtor tests.
